refactor: 重构首页代码结构
