+++ frontend/src/views/home/constants/chestConfig.js
@@ -1,0 +1,70 @@
+// 宝箱配置
+
+// 宝箱楼层配置
+export const chestLevels = [10, 20, 50, 60, 70, 80, 90, 110, 120]
+
+// 宝箱物品配置（根据楼层）
+export const chestItems = {
+  10: ['皮靴', '工作靴', '木剑', '铁制短剑', '疾风利剑', '股骨'],
+  20: ['钢制轻剑', '木棒', '精灵之刃', '光辉戒指', '磁铁戒指'],
+  50: ['冻土靴', '热能靴', '战靴', '镀银军刀', '海盗剑'],
+  60: ['水晶匕首', '弯刀', '铁刃', '飞贼之胫', '木锤'],
+  70: ['圣堂之刃'],
+  80: ['蹈火者靴', '黑暗之靴', '双刃大剑', '圣堂之刃', '长柄锤', '暗影匕首'],
+  90: ['黑曜石之刃', '淬火阔剑', '蛇形邪剑', '骨剑', '骨化剑'],
+  110: ['太空之靴', '水晶鞋', '钢刀', '巨锤'],
+  120: ['骷髅钥匙']
+}
+
+// 宝箱预设模板
+export const chestPresets = [
+  {
+    name: '20层磁铁戒指',
+    description: '获取20层的磁铁戒指',
+    mode: 'ALL',
+    rules: [
+      { type: 'atom', level: 20, item: '磁铁戒指' }
+    ]
+  },
+  {
+    name: '80+110层双套装',
+    description: '(80层=长柄锤 且 110层=太空之靴) 或 (80层=蹈火者靴 且 110层=巨锤)',
+    mode: 'ALL',
+    rules: [
+      {
+        type: 'or_group',
+        items: [
+          [
+            { level: 80, item: '长柄锤' },
+            { level: 110, item: '太空之靴' }
+          ],
+          [
+            { level: 80, item: '蹈火者靴' },
+            { level: 110, item: '巨锤' }
+          ]
+        ]
+      }
+    ]
+  },
+  {
+    name: '完整配置示例',
+    description: '20层=磁铁戒指 且 (80+110层双套装)',
+    mode: 'ALL',
+    rules: [
+      { type: 'atom', level: 20, item: '磁铁戒指' },
+      {
+        type: 'or_group',
+        items: [
+          [
+            { level: 80, item: '长柄锤' },
+            { level: 110, item: '太空之靴' }
+          ],
+          [
+            { level: 80, item: '蹈火者靴' },
+            { level: 110, item: '巨锤' }
+          ]
+        ]
+      }
+    ]
+  }
+]
