+++ frontend/src/views/home/constants/weatherConfig.js
@@ -1,0 +1,10 @@
+// 天气配置
+
+export const weatherOptions = [
+  { label: '雨天', value: 'Rain' },
+  { label: '雷雨天', value: 'Storm' },
+  { label: '绿雨天', value: 'Green Rain' },
+  { label: '大风天', value: 'Wind' },
+  { label: '雪天', value: 'Snow' },
+  { label: '晴天', value: 'Sun' },
+]
