+++ frontend/src/views/weather-forecast/weather-forecast.vue
